attendee stuff
